Initial commit
Given a photo with a file name in a ceryain format, this program creates HTML pages, with metadata extracted from the file name. 

This command kicks off the Python program, NewPhoto-Add.py, to take all the photos in this directory (/XXX//NewPhoto-Add/Delivery), and make accompanying photo pages from them, using the metadata in the file name. Multiple photos can be added. For now, the HTML pages are delivered in the same folder. Note Delivery must have full write capability for this to work.

Note, this is part of a larger processs that runs automatically, each Monday morning, that process all the photos in the delivery folder than places them in the current live photo directly (Currently 2015).

For best results use this format for labelling photos:
YYMM-PhotoType-AuthorOrEventOfPhotoType-NameOrSpecificOfPhotoType-00(if multiples)... So From general to specific, really.
Shorthand for PhotoType:

                "SA": "Street Art",
                "PP" : "People",
                "RR" : "Railroad",
				"MU" : "Music",
                "FD" : "Food and Drink",
                "FA" : "Fine Art",
                "VE" : "Vehicles",
                "PL" : "Place",
                "EV" : "Event",
                "NA" : "Nature",
                "Tech" : "Tech",
				"CL" : "Clothing",
				"Misc" : "Miscellanious",
				"Folk" : "Folk"

Others can be added to the NewPhoto-Add.py source code.